requirements.txt
crispy-bootstrap5==2023.10
Django==4.2.5
django-credit-cards==0.4.1
django-crispy-forms==2.1
django-localflavor==4.0
django-phonenumber-field==7.2.0
django-phonenumbers
Pillow==9.0.1
pynvim==0.4.3
